cyprus-price-comparison/requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
urllib3==2.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
urllib-robotparser==1.0.0
aiohttp==3.9.1
